% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\name{bank_getList}
\alias{bank_getList}
\title{查询银行账户信息}
\usage{
bank_getList(token = "34A4B98D-AD90-4691-BDCC-63A86C46A22F")
}
\arguments{
\item{token}{口令}
}
\value{
返回值
}
\description{
查询银行账户信息
}
\examples{
bank_getList()
}
