% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.R
\name{company_getList}
\alias{company_getList}
\title{返回公司列表}
\usage{
company_getList(dms_token = "34A4B98D-AD90-4691-BDCC-63A86C46A22F")
}
\arguments{
\item{dms_token}{口令}
}
\value{
返回公司列表
}
\description{
返回公司列表
}
\examples{
company_getList()
}
