% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.R
\name{company_isGroup}
\alias{company_isGroup}
\title{判断公司是否集团}
\usage{
company_isGroup(
  dms_token = "34A4B98D-AD90-4691-BDCC-63A86C46A22F",
  FCompany = "赛普集团"
)
}
\arguments{
\item{dms_token}{口令}

\item{FCompany}{公司名称}
}
\value{
返回公司列表
}
\description{
判断公司是否集团
}
\examples{
company_isGroup()
}
