% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyRpt.R
\name{dailyRpt_query}
\alias{dailyRpt_query}
\title{返回公司列表}
\usage{
dailyRpt_query(
  dms_token = "34A4B98D-AD90-4691-BDCC-63A86C46A22F",
  FDate = "2022-12-01"
)
}
\arguments{
\item{dms_token}{口令}

\item{FDate}{日期}
}
\value{
返回公司列表
}
\description{
返回公司列表
}
\examples{
company_getList()
}
