% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_query}
\alias{transaction_query}
\title{返回交易明细信息}
\usage{
transaction_query(
  dms_token = "34A4B98D-AD90-4691-BDCC-63A86C46A22F",
  FDate = "2022-12-01",
  FAccountNO = "467677162516"
)
}
\arguments{
\item{dms_token}{口令}

\item{FDate}{日期}

\item{FAccountNO}{银行账号}
}
\value{
返回公司列表
}
\description{
返回交易明细信息
}
\examples{
transaction_query()
}
