% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction_query_server}
\alias{transaction_query_server}
\title{账户查询函数}
\usage{
transaction_query_server(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
账户查询函数
}
\examples{
transaction_query_server()
}
